Task 1 (Web Scraping and Data Analysis)
In this first task we have done web scraping to the Skytrax website, to collect customer reviews of the company with subsequent cleaning and filtering. 
Finally we have taken the cleaned data to analyze the customer reviews and get clear ideas about the needs of the company for its customers.